+++ src/main/java/com/example/KTB_4WEEK/dto/response/post/IncreaseLikeResponseDto.java
@@ -10,11 +10,11 @@ public class IncreaseLikeResponseDto {
     @Schema(description = "좋아요 수 증가 대상 게시물 PK", example = "1")
     private long id;
     @Schema(description = "게시물의 총 좋아요 개수", example = "100")
-    private int like;
+    private long like;
     @Schema(description = "게시물 최근 업데이트 시간", example = "2025-10-10 20:00:00")
     private String updateAt;
 
-    public IncreaseLikeResponseDto(long id, int like, LocalDateTime updateAt) {
+    public IncreaseLikeResponseDto(long id, long like, LocalDateTime updateAt) {
         this.id = id;
         this.like = like;
         this.updateAt = updateAt.format(DateTimeFormatter
@@ -25,7 +25,7 @@ public class IncreaseLikeResponseDto {
         return id;
     }
 
-    public int getLike() {
+    public long getLike() {
         return like;
     }
